requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
redis==5.0.0
# Database drivers 
PyMySQL==1.1.0
# Server / workers
gunicorn==21.2.0
celery==5.4.0     

# Utilities
Pillow>=10.1.0
pyotp==2.9.0
qrcode==7.4.2
stripe==6.6.0
requests==2.31.0
